Eliminate dependency on ReginsterMetaTypes in AudioInjectorOptions,
by moving the script value conversion functions into
AudioScriptingInterface.
